add aziotctl system stop (#176)
Super simple change to add stop to the system commands
