+++ aziotctl/aziotctl-common/src/restart.rs
@@ -1,27 +1,14 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-use std::io::{self, Write};
 use std::process::Command;
 
 use anyhow::{Context, Result};
 
-use crate::ServiceDefinition;
+use crate::{print_command_error, stop, ServiceDefinition};
 
 pub fn restart(services: &[&ServiceDefinition]) -> Result<()> {
     // stop all services
-    for service in services.iter().map(|s| s.service) {
-        print!("Stopping {}...", service);
-        let result = Command::new("systemctl")
-            .args(&["stop", service])
-            .output()
-            .context("Failed to call systemctl stop")?;
-
-        if result.status.success() {
-            println!("Stopped!");
-        } else {
-            print_command_error(&result);
-        }
-    }
+    stop(services)?;
 
     // start all sockets
     for socket in services.iter().flat_map(|s| s.sockets) {
@@ -47,15 +34,4 @@ fn start(name: &str) -> Result<()> {
     }
 
     Ok(())
-}
-
-fn print_command_error(result: &std::process::Output) {
-    eprintln!("systemctl exited with non-zero status code.");
-    eprintln!("stdout:");
-    eprintln!("=======");
-    io::stdout().write_all(&result.stdout).unwrap();
-    eprintln!("stderr:");
-    eprintln!("=======");
-    io::stdout().write_all(&result.stderr).unwrap();
-    eprintln!();
 }
