+++ aziotctl/src/system.rs
@@ -5,11 +5,14 @@ use std::ffi::{OsStr, OsString};
 use anyhow::{anyhow, Result};
 use structopt::StructOpt;
 
-use aziotctl_common::{get_status, get_system_logs, restart, set_log_level, SERVICE_DEFINITIONS};
+use aziotctl_common::{
+    get_status, get_system_logs, restart, set_log_level, stop, SERVICE_DEFINITIONS,
+};
 
 #[derive(StructOpt)]
 pub enum Options {
     Restart(RestartOptions),
+    Stop(StopOptions),
     Status(StatusOptions),
     Logs(LogsOptions),
     SetLogLevel(LogLevelOptions),
@@ -19,6 +22,10 @@ pub enum Options {
 #[derive(StructOpt)]
 #[structopt(about = "Restart the services")]
 pub struct RestartOptions {}
+
+#[derive(StructOpt)]
+#[structopt(about = "Stop the services")]
+pub struct StopOptions {}
 
 #[derive(StructOpt)]
 #[structopt(about = "Report the status of system")]
@@ -54,6 +61,7 @@ pub struct ReprovisionOptions {
 pub async fn system(options: Options) -> Result<()> {
     match options {
         Options::Restart(_) => restart(SERVICE_DEFINITIONS),
+        Options::Stop(_) => stop(SERVICE_DEFINITIONS),
         Options::Status(_) => get_status(SERVICE_DEFINITIONS),
         Options::Logs(opts) => logs(&opts),
         Options::SetLogLevel(opts) => set_log_level(SERVICE_DEFINITIONS, opts.log_level),
